requirements.txt
python-dotenv
requests
PyGithub